+++ vizdet/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "0.1.0"
 
-from .bboxes import BBoxes  # noqa: F401
+from .bboxes import BBoxes, ColorMode  # noqa: F401
 from .font import Font  # noqa: F401
 from .infobox import InfoBox  # noqa: F401
 from .label import Label  # noqa: F401